@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
+
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
 import { AppRoutingModule } from './app-routing.module';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
 import { ActionnaireComponent } from './Historique/actionnaire/actionnaire.component';
@@ -22,62 +23,52 @@ import { provideAnimationsAsync } from '@angular/platform-browser/animations/asy
 import { NatureCompteTitreComponent } from './Historique/nature-compte-titre/nature-compte-titre.component';
 import { TitreComponent } from './Historique/titre/titre.component'; 
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
-import { ImportComponent } from './Referentiel/import/import.component';
 import { MatRadioModule } from '@angular/material/radio';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatNativeDateModule } from '@angular/material/core';
+import { ImportComponent } from './Referentiel/import/import.component';
 
 
 
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    ActionnaireComponent,
-    TitreComponent,
-    NatureAvoirComponent,
-    TypeOperationComponent,
-    TeneurCompteComponent,
-    NatureAvoirComponent,
-    NatureCompteTitreComponent,
-    TitreComponent,
-    ImportComponent,
-    
-
-
-
-   ],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    MatTableModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatPaginatorModule,
-    MatSortModule,
-    RouterModule,
-    AppRoutingModule,
-    HttpClientModule,
-    MatSelectModule,
-    FormsModule,
-    MatIconModule,
-    FormsModule,
-    ReactiveFormsModule,
-    MatIconModule,
-    MatRadioModule,
-    MatButtonModule,
-    MatIconModule,
-    AppRoutingModule,
-    MatDatepickerModule,
-    MatNativeDateModule,
-    
-  ],
-  providers: [
-    provideAnimationsAsync(),
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        ActionnaireComponent,
+        TitreComponent,
+        NatureAvoirComponent,
+        TypeOperationComponent,
+        TeneurCompteComponent,
+        NatureAvoirComponent,
+        NatureCompteTitreComponent,
+        TitreComponent,
+        ImportComponent,
+        
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        BrowserAnimationsModule,
+        MatTableModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatPaginatorModule,
+        MatSortModule,
+        RouterModule,
+        AppRoutingModule,
+        MatSelectModule,
+        FormsModule,
+        MatIconModule,
+        FormsModule,
+        ReactiveFormsModule,
+        MatIconModule,
+        MatRadioModule,
+        MatButtonModule,
+        MatIconModule,
+        AppRoutingModule,
+        MatDatepickerModule,
+        MatNativeDateModule], providers: [
+        provideAnimationsAsync(),
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule { 
   
 }
